Updating 'header animation' section for v1 item file format
The EEs are built with One Pixel Production's expanded animations, resulting in valid values for this field not being documented
for the EEs.

Also documented the minor differences between the BG/IWD and BG2 animations, specifically the difference in helmet appearance and the
availability of the S3 (katana)/SC (scimitar) animations. While the shield cosmetics also vary, I didn't document this as they're
internally consistent within the games.